+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## [unreleased]
 
+ * Portal: Added the attribute *data-mr-app-name* to the default App wrapper to simplify end-2-end testing
  * Portal: The SSR route of Remote Apps will receive now also the path and the query parameters of the original request (fixes #102).
    The body of the POST request looks like this now:
    ```ts